fix(feature-selection): widen n*n normalisers in distance-correlation, kernel and SURF selectors
CodeQL cs/loss-of-precision: n*n (and n1*n2, n*|near hits|) normalisers were int products.
They overflow at n >= 46,341 samples. For the selectors that also allocate an n x n double[,]
this is a large-memory edge (>= 17 GB, still allocatable on 64-bit .NET); MMD, SURF/MultiSURF
and SeasonalityFS (series length, e.g. minute bars) reach it without such an allocation.

+++ src/Preprocessing/FeatureSelection/Filter/DistanceCorrelation.cs
@@ -159,7 +159,7 @@ public class DistanceCorrelation<T> : TransformerBase<T, Matrix<T>, Matrix<T>>
             rowMeans[i] /= n;
             colMeans[i] /= n;
         }
-        grandMean /= (n * n);
+        grandMean /= ((double)n * n);
 
         // Double centering
         for (int i = 0; i < n; i++)
@@ -176,7 +176,7 @@ public class DistanceCorrelation<T> : TransformerBase<T, Matrix<T>, Matrix<T>>
             for (int k = 0; k < n; k++)
                 sum += a[i, k] * b[i, k];
 
-        return sum / (n * n);
+        return sum / ((double)n * n);
     }
 
     public Matrix<T> FitTransform(Matrix<T> data, Vector<T> target)
